cookie/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
python-dotenv==1.0.1
Flask-Login==0.6.3